requirement.txt
sentence-transformers
faiss-cpu
transformers
torch
tqdm
huggingface-hub
numpy
streamlit